Moved internal bot files to a subdirectory.
