+++ .gitignore
@@ -3,16 +3,17 @@
 dk
 __pycache__
 tempCodeRunnerFile.py
-data/carts/limbo/*.json
-data/carts/*.json
-!data/carts/0.json
-!data/carts/1.json
-data/wallets/*.json
-!data/wallets/0.json
-!data/wallets/1.json
-!data/wallets/91.json
-data/transactions/*.json
-!data/transactions/0.json
-!data/transactions/1.json
-!data/transactions/91.json
-trials/
+src/tempCodeRunnerFile.py
+src/data/carts/limbo/*.json
+src/data/carts/*.json
+!srcdata/carts/0.json
+!src/data/carts/1.json
+src/data/wallets/*.json
+!src/data/wallets/0.json
+!src/data/wallets/1.json
+!src/data/wallets/91.json
+src/data/transactions/*.json
+!src/data/transactions/0.json
+!src/data/transactions/1.json
+!src/data/transactions/91.json
+src/trials/
