+++ main.py
@@ -44,7 +44,7 @@ def path_finder(path: str):
 
 # Path to bot.py
 botfile = os.path.dirname(os.path.abspath(
-    __file__)) + "/bot.py"
+    __file__)) + "/src/bot.py"
 botfile = path_finder(botfile)
 botfile = '"' + botfile + '"'
