+++ src/data/commands/help.json
